fix: restore bundled runtime dependency provisioning
