+++ test/scripts/stage-bundled-plugin-runtime-deps.test.ts
@@ -2,127 +2,7 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
-import {
-  resolveNpmRunner,
-  stageBundledPluginRuntimeDeps,
-} from "../../scripts/stage-bundled-plugin-runtime-deps.mjs";
-
-describe("resolveNpmRunner", () => {
-  it("anchors npm staging to the active node toolchain when npm-cli.js exists", () => {
-    const execPath = "/Users/test/.nodenv/versions/24.13.0/bin/node";
-    const expectedNpmCliPath = path.posix.resolve(
-      path.posix.dirname(execPath),
-      "../lib/node_modules/npm/bin/npm-cli.js",
-    );
-
-    const runner = resolveNpmRunner({
-      execPath,
-      env: {},
-      existsSync: (candidate: string) => candidate === expectedNpmCliPath,
-      platform: "darwin",
-    });
-
-    expect(runner).toEqual({
-      command: execPath,
-      args: [expectedNpmCliPath],
-      shell: false,
-    });
-  });
-
-  it("anchors Windows npm staging to the adjacent npm-cli.js without a shell", () => {
-    const execPath = "C:\\nodejs\\node.exe";
-    const expectedNpmCliPath = path.win32.resolve(
-      path.win32.dirname(execPath),
-      "node_modules/npm/bin/npm-cli.js",
-    );
-
-    const runner = resolveNpmRunner({
-      execPath,
-      env: {},
-      existsSync: (candidate: string) => candidate === expectedNpmCliPath,
-      platform: "win32",
-    });
-
-    expect(runner).toEqual({
-      command: execPath,
-      args: [expectedNpmCliPath],
-      shell: false,
-    });
-  });
-
-  it("uses an adjacent npm.exe on Windows without a shell", () => {
-    const execPath = "C:\\nodejs\\node.exe";
-    const expectedNpmExePath = path.win32.resolve(path.win32.dirname(execPath), "npm.exe");
-
-    const runner = resolveNpmRunner({
-      execPath,
-      env: {},
-      existsSync: (candidate: string) => candidate === expectedNpmExePath,
-      npmArgs: ["install", "--silent"],
-      platform: "win32",
-    });
-
-    expect(runner).toEqual({
-      command: expectedNpmExePath,
-      args: ["install", "--silent"],
-      shell: false,
-    });
-  });
-
-  it("wraps an adjacent npm.cmd via cmd.exe without enabling shell mode", () => {
-    const execPath = "C:\\nodejs\\node.exe";
-    const npmCmdPath = path.win32.resolve(path.win32.dirname(execPath), "npm.cmd");
-
-    const runner = resolveNpmRunner({
-      comSpec: "C:\\Windows\\System32\\cmd.exe",
-      execPath,
-      env: {},
-      existsSync: (candidate: string) => candidate === npmCmdPath,
-      npmArgs: ["install", "--omit=dev"],
-      platform: "win32",
-    });
-
-    expect(runner).toEqual({
-      command: "C:\\Windows\\System32\\cmd.exe",
-      args: ["/d", "/s", "/c", `${npmCmdPath} install --omit=dev`],
-      shell: false,
-      windowsVerbatimArguments: true,
-    });
-  });
-
-  it("prefixes PATH with the active node dir when falling back to bare npm", () => {
-    expect(
-      resolveNpmRunner({
-        execPath: "/tmp/node",
-        env: {
-          PATH: "/usr/bin:/bin",
-        },
-        existsSync: () => false,
-        platform: "linux",
-      }),
-    ).toEqual({
-      command: "npm",
-      args: [],
-      shell: false,
-      env: {
-        PATH: `/tmp${path.delimiter}/usr/bin:/bin`,
-      },
-    });
-  });
-
-  it("fails closed on Windows when no toolchain-local npm CLI exists", () => {
-    expect(() =>
-      resolveNpmRunner({
-        execPath: "C:\\node\\node.exe",
-        env: {
-          Path: "C:\\Windows\\System32",
-        },
-        existsSync: () => false,
-        platform: "win32",
-      }),
-    ).toThrow("OpenClaw refuses to shell out to bare npm on Windows");
-  });
-});
+import { stageBundledPluginRuntimeDeps } from "../../scripts/stage-bundled-plugin-runtime-deps.mjs";
 
 describe("stageBundledPluginRuntimeDeps", () => {
   function createBundledPluginFixture(params: {
@@ -227,6 +107,28 @@ describe("stageBundledPluginRuntimeDeps", () => {
 
     expect(installCount).toBe(2);
     expect(fs.readFileSync(path.join(pluginDir, "node_modules", "marker.txt"), "utf8")).toBe("2\n");
+  });
+
+  it("stages runtime deps from the root node_modules when already installed", () => {
+    const { pluginDir, repoRoot } = createBundledPluginFixture({
+      packageJson: {
+        name: "@openclaw/fixture-plugin",
+        version: "1.0.0",
+        dependencies: { "left-pad": "1.3.0" },
+        openclaw: { bundle: { stageRuntimeDependencies: true } },
+      },
+    });
+    const rootDepDir = path.join(repoRoot, "node_modules", "left-pad");
+    fs.mkdirSync(rootDepDir, { recursive: true });
+    fs.writeFileSync(path.join(rootDepDir, "package.json"), '{ "name": "left-pad" }\n', "utf8");
+    fs.writeFileSync(path.join(rootDepDir, "index.js"), "module.exports = 1;\n", "utf8");
+
+    stageBundledPluginRuntimeDeps({ cwd: repoRoot });
+
+    expect(
+      fs.readFileSync(path.join(pluginDir, "node_modules", "left-pad", "index.js"), "utf8"),
+    ).toBe("module.exports = 1;\n");
+    expect(fs.existsSync(path.join(pluginDir, ".openclaw-runtime-deps-stamp.json"))).toBe(true);
   });
 
   it("retries transient runtime dependency staging failures before surfacing an error", () => {
